Changed m_seqRefreshRate and associated usage from an int to a float. This handles when a sequence file has an uncommon Step Time. This causes the result of "1000 / Step Time" to not be an integer. Discovered with a Step Time of 17ms which caused the audio and lights to not stay in sync. At the end of the audio, the lights continued for several seconds before completing.

+++ src/Sequence.h
@@ -94,7 +94,7 @@ class Sequence {
     int           m_seqStepTime;
 	int           m_seqSingleStep;
 	int           m_seqSingleStepBack;
-	int           m_seqRefreshRate;
+	float           m_seqRefreshRate;
 	int           m_seqControlRawIDs;
 	char          m_seqLastControlMajor;
 	char          m_seqLastControlMinor;
